adding setting story and link it with home fragmet

+++ app/src/main/java/com/example/weatherapp/models/Hourly.kt
@@ -8,17 +8,17 @@ data class Hourly (
   @SerializedName("dt"         ) var dt         : Long?               = null,
   @SerializedName("temp"       ) var temp       : Double?            = null,
   @SerializedName("feels_like" ) var feelsLike  : Double?            = null,
-  @SerializedName("pressure"   ) var pressure   : Int?               = null,
-  @SerializedName("humidity"   ) var humidity   : Int?               = null,
+  @SerializedName("pressure"   ) var pressure   : Double?               = null,
+  @SerializedName("humidity"   ) var humidity   : Double?               = null,
   @SerializedName("dew_point"  ) var dewPoint   : Double?            = null,
   @SerializedName("uvi"        ) var uvi        : Double?               = null,
-  @SerializedName("clouds"     ) var clouds     : Int?               = null,
-  @SerializedName("visibility" ) var visibility : Int?               = null,
+  @SerializedName("clouds"     ) var clouds     : Double?               = null,
+  @SerializedName("visibility" ) var visibility : Double?               = null,
   @SerializedName("wind_speed" ) var windSpeed  : Double?            = null,
-  @SerializedName("wind_deg"   ) var windDeg    : Int?               = null,
+  @SerializedName("wind_deg"   ) var windDeg    : Double?               = null,
   @SerializedName("wind_gust"  ) var windGust   : Double?            = null,
   @SerializedName("weather"    ) var weather    : ArrayList<Weather> = arrayListOf(),
-  @SerializedName("pop"        ) var pop        : Int?               = null,
+  @SerializedName("pop"        ) var pop        : Double?               = null,
   @SerializedName("rain"       ) var rain       : Rain?              = Rain()
 
 )
